Changes to way this is launched, to support dev or prod mode.
The default mode is now prod when launched with:

    docker-compose up

To launch in dev mode:

    docker-compose -f docker-compose.yml -f docker-compose-dev.yml up

The latter method basically enables flask's live reloader, and dev mode.
This makes it easier to change code as the service is running; previously
you had to issue a `--build` command to account for code changes.

I may add an nginx service to the former method in future.

I also changed the `processor/Dockerfile`, to move `./model_data` to
`/model_data` within the container.  This allows for volume mounts to be used
in dev mode, without making the models invisible in the continer.